fix : hotel type error
Former-commit-id: b2af6772230b08c841f12874a58d35f15c57e69a

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/hotel/dto/request/AttractionCoordinateRequestDto.java
@@ -11,6 +11,6 @@ import java.math.BigDecimal;
 public class AttractionCoordinateRequestDto {
     private Double latitude;
     private Double longitude;
-    private Long distance;
+    private Integer distance;
     private String sorted;
 }
